+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Add a way to specify a base recipe with `from: base-recipe`. More information in the docs
 - Add `--no-color` option that disables colored output
 - Add `--output-dir` flag to `build` subcommand to override output directory for this build
+- Add `no_color` config option to disable color output globally in configuration
 
 # 0.8.0
 - Dependencies defined for default images like `pkger-deb` or `pkger-rpm` will be used for custom images using the same output target. This means that all images used for building RPM will use dependencies defined for `pkger-rpm` - cleaner recipes ;)